refactor : split AddQuickMemoButton from Header and add useState hook for rotating css
